+++ src/components/StopwatchButton.tsx
